Real GUI/tests/requirements.txt
pytest
pytest-timeout
msgpack-rpc-python
paramiko
numpy
matplotlib
ConfigParser
